Добавлено ограничение на проставление рейтинга одному и тому же посту одним и тем же юзером

+++ models/userPostRating.go
@@ -1,7 +1,7 @@
 package models
 
 type UserPostRating struct {
-	UserId int    `json:"userId"`
+	UserId int64  `json:"userId"`
 	PostId string `json:"postId"`
 	Oper   rune   `json:"oper"`
 }
